Switched to float from double

+++ CPP Google for Education/Chirper/chirper.cpp
@@ -4,14 +4,14 @@ char yn ='y';
 int main()
 {
   int chirps{0};
-  double temp{10.0};
+  float temp{10.0};
   while(yn=='y')
   {
     
     //Get chirps
     std::cout<<"Enter number of chirps: ";
     std::cin>>chirps;
-    temp=((double) chirps+40)/4;
+    temp=((float) chirps+40)/4;
     std::cout<<"The temperature is "<<std::setprecision(1)<<std::fixed<<temp<<" degrees.\n";
     
     yn=AskRepeat("Check another?",yn); //replace with ask phrase
